Implemented vertical victory

+++ src/entities/coordinates.rs
@@ -1,10 +1,10 @@
 pub struct Coordinates {
-    pub x: u32,
-    pub y: u32
+    pub x: usize,
+    pub y: usize
 }
 
 impl Coordinates {
-    pub fn from(x: u32, y: u32) -> Coordinates {
+    pub fn from(x: usize, y: usize) -> Coordinates {
         Coordinates {x, y}
     }
 }
